Generated simulated reads
Script for creating simulated reads using polyester. A total of 60 genes from
chr22 were chosen, 24 with only 1 transcript and 36 with 2. For the first 24, 12
were randomly chosen to be DE (2 high, 2 up per group; 3 groups in total) and 12
are not DE. For the 36 with 2 transcripts, 12 are DE on both transcripts, 12 are
DE in only 1 transcript, and 12 are not DE.