Coverage plots with F-stats for false negative/positives, segments vs DERs
The one false negative from the _oneDE_ case seems to be due to low coverage and
could be mitigated by a lower F-stat cutoff.

The apparent false positives from the _oneDE_ case can be fairly complex. In
particular some overlap two exons on one strand and a third (or more) on the
opposite strand. It's good to see that exons not overlapping other exons don't
contain DERs.

This new version also includes an exploration of the performance at the exonic
segment level against DERs. There are 37 false negatives and 0 false positives
when using the DERs with sig FWER. The majority of the false negatives come from
cases where one of two transcripts of the gene are set to be DE. Only one has
no coverage data. In some cases, lowering the F-stat cutoff could help, just
like it was seen at the transcript level.