Completed figure 2
Made tracks for coverage per, mean coverage with regions from Region
Matrix showcased, genes, and transcripts. Also made a track for the F-stats
but did not include it on the figure.

Note that the transcript track was more complicated than in figure 1 due to
the presence of transcripts from both strands.