Re-made compareVsPNAS.html in two parts: R in the cluster, pandoc in laptop
The cluster hanged twice for 16+ hrs on the pandoc part, so I just finished it
in 4 minutes in my laptop. I don't understand why the cluster is failing the
pandoc step, but that's the current situation. Note that the pandoc versions
are different as it's not easy to find an updated version to install in the
cluster.

In timing.Rmd I'll have to manually modify the time spent in this job.