Using R 3.1.x, load cov and get region matrix for gtex data
From previous tests, it does look like the memory reporting changed between
R 3.1.x and R 3.2.0 (same as 3.2.x). I'll ask Bioc-devel about this, but in the
meantime, the easiest option is to use R 3.1.x.

I modified the filtering at step 1 so the coverage data gets normalized before
filtering, since we wanted to do this for the gtex samples.